# docs/swagger.yaml
basePath: /
definitions:
...
  report.AttendeeListEnvelope:
    properties:
      data:
        items:
          $ref: '#/definitions/report.AttendeeResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  report.AttendeeResponse:
    properties:
      admittedCount:
        description: AdmittedCount is how many of this line's Quantity tickets have
          entered.
        example: 2
        type: integer
      ageYears:
        description: AgeYears is the buyer's age AT PURCHASE, 0 when not informed.
        example: 31
        type: integer
      city:
        example: Natal
        type: string
      documentMask:
        example: 529***25
        type: string
      email:
        example: maria@exemplo.com.br
        type: string
      gender:
        example: female
        type: string
      name:
        example: Maria Souza
        type: string
      netCents:
        example: 48000
        type: integer
      orderId:
        example: ord_4a1b7c
        type: string
      purchasedAt:
        type: string
      quantity:
        example: 2
        type: integer
      status:
        example: paid
        type: string
      ticketId:
        example: tkt_9f
        type: string
      ticketTitle:
        example: Pista
        type: string
      uf:
        example: RN
        type: string
      unitPriceCents:
        example: 24000
        type: integer
    type: object
  report.DayResponse:
    properties:
      day:
        description: Day is YYYY-MM-DD in UTC, the day the money LANDED.
        example: "2026-07-26"
        type: string
      netCents:
        example: 15920000
        type: integer
      orders:
        example: 612
        type: integer
      tickets:
        example: 1592
        type: integer
    type: object
  report.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
    type: object
  report.SalesEnvelope:
    properties:
      data:
        $ref: '#/definitions/report.SalesResponse'
    type: object
  report.SalesResponse:
    properties:
      byAge:
        items:
          $ref: '#/definitions/report.SliceResponse'
        type: array
      byCity:
        items:
          $ref: '#/definitions/report.SliceResponse'
        type: array
      byDay:
        items:
          $ref: '#/definitions/report.DayResponse'
        type: array
      byGender:
        items:
          $ref: '#/definitions/report.SliceResponse'
        type: array
      byTier:
        items:
          $ref: '#/definitions/report.SliceResponse'
        type: array
      byUf:
        items:
          $ref: '#/definitions/report.SliceResponse'
        type: array
      eventId:
        example: evt_88
        type: string
      totals:
        $ref: '#/definitions/report.TotalsResponse'
    type: object
  report.SliceResponse:
    properties:
      key:
        example: SP
        type: string
      label:
        example: Pista
        type: string
      netCents:
        description: |-
          NetCents is the organiser's earnings for this slice. It is the only
          money field here; see the money note in domain/report.
        example: 9370000
        type: integer
      orders:
        description: |-
          Orders is how many purchases; Tickets is how many admissions. One order
          for eight tickets is one buyer and eight seats, and a table showing only
          one of them is wrong for half the decisions made from it.
        example: 412
        type: integer
      tickets:
        example: 937
        type: integer
    type: object
  report.TotalsResponse:
    properties:
      averageOrderCents:
        description: |-
          AverageOrderCents and AverageTicketCents are the ticket médio, asked the
          two ways an organiser means it: what one buyer spends in a go, and what
          one admission is worth. They differ whenever anybody buys for a group,
          and only the second is evidence for repricing a tier.

          Sent rather than left to the client to divide, so the API, the CSV and
          every screen round the same way.
        example: 28750
        type: integer
      averageTicketCents:
        example: 10000
        type: integer
      buyers:
        description: 'Buyers is DISTINCT accounts: how many PEOPLE, not how many purchases.'
        example: 11902
        type: integer
      netCents:
        description: |-
          NetCents is what the organiser earned: the sum of the face values they
          priced. RefundedCents is at face value too.

          Neither the gross the buyer paid nor our commission is sent, and not
          because they are filtered out here: the query never selects them. See
          the money note in domain/report.
        example: 357350000
        type: integer
      orders:
        example: 12430
        type: integer
      refundedCents:
        example: 2750000
        type: integer
      refundedOrders:
        example: 87
        type: integer
      tickets:
        example: 35735
        type: integer
    type: object
...
info:
  contact: {}
  description: API base para autenticação e gestão de tickets, organizada com Clean
    Architecture.
  title: Vozkot Tickets API
  version: "1.0"
paths:
...
  /api/v1/events/{id}/attendees:
    get:
      description: 'A lista de participantes, uma linha por lote de cada pedido. Traz
        nome, e-mail, o documento MASCARADO, e o que o comprador informou de sexo,
        idade, cidade e estado. O documento completo, a data de nascimento e o telefone
        nunca são devolvidos: o organizador precisa conferir quem chega na porta,
        não se passar por ela. Por padrão lista apenas pedidos pagos; use `status=all`
        para incluir os estornados.'
      parameters:
      - description: ID do evento
        in: path
        name: id
        required: true
        type: string
      - description: Filtrar por lote
        in: query
        name: ticketId
        type: string
      - description: 'Situação do pedido (padrão: paid)'
        enum:
        - paid
        - refunded
        - all
        in: query
        name: status
        type: string
      - description: Buscar por nome ou e-mail
        in: query
        name: q
        type: string
      - description: Itens por página (padrão 50, máximo 200)
        in: query
        name: limit
        type: integer
      - description: Deslocamento da paginação
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/report.AttendeeListEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/report.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/report.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/report.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar quem comprou
      tags:
      - Relatórios
  /api/v1/events/{id}/attendees.csv:
    get:
      description: Baixa a lista de participantes como uma planilha. Aceita os mesmos
        filtros da listagem. O arquivo é UTF-8 com BOM e separado por ponto e vírgula,
        que é o que o Excel em português abre corretamente sem nenhum passo de importação;
        os valores vão em reais com vírgula decimal. As linhas são transmitidas em
        lotes conforme são lidas, então exportar um estádio inteiro não carrega tudo
        na memória.
      parameters:
      - description: ID do evento
        in: path
        name: id
        required: true
        type: string
      - description: Filtrar por lote
        in: query
        name: ticketId
        type: string
      - description: 'Situação do pedido (padrão: paid)'
        enum:
        - paid
        - refunded
        - all
        in: query
        name: status
        type: string
      - description: Buscar por nome ou e-mail
        in: query
        name: q
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: CSV
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/report.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/report.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/report.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Exportar quem comprou (CSV)
      tags:
      - Relatórios
...
  /api/v1/events/{id}/report:
    get:
      description: 'Os números do evento para o organizador: totais, e a divisão por
        sexo, faixa etária, estado, cidade, lote e dia. Ninguém é identificado aqui:
        são contagens sobre os dados que o comprador informou no cadastro, congelados
        no momento da compra. Todo valor aqui é `netCents`: o que o organizador recebe,
        ou seja, o valor de face que ele definiu. A taxa de serviço da plataforma
        e o total pago pelo comprador não são retornados neste relatório. Restrito
        ao organizador do evento e a administradores.'
      parameters:
      - description: ID do evento
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/report.SalesEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/report.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/report.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/report.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Relatório de vendas do evento
      tags:
      - Relatórios
...
  /api/v1/organiser/report:
    get:
      description: 'Os mesmos números do relatório de evento, somados sobre TODOS
        os eventos do organizador: totais, ticket médio, e a divisão por sexo, faixa
        etária, estado, cidade, lote e dia. Ninguém é identificado: são contagens
        sobre o que o comprador informou no cadastro, congelado no momento da compra.
        O escopo vem da sessão: não existe parâmetro de organizador, portanto não
        há o que adulterar para ler a carteira de outra pessoa.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/report.SalesEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/report.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Relatório geral do organizador
      tags:
      - Relatórios
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Use o formato: Bearer {token}. Navegadores também podem autenticar
      por cookies httpOnly.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"